Try to fix tests
