+++ src/test/java/org/folio/rest/impl/CustomFieldTestBase.java
@@ -11,16 +11,12 @@ import static org.folio.test.util.TestUtil.readFile;
 import static org.folio.test.util.TestUtil.toJson;
 
 import java.io.IOException;
-import java.lang.reflect.Field;
 import java.net.URISyntaxException;
 import java.util.Arrays;
-import java.util.List;
 
 import com.github.tomakehurst.wiremock.matching.EqualToPattern;
 import io.restassured.http.Header;
 import io.vertx.core.json.Json;
-import org.folio.event.KafkaConfigSingleton;
-import org.folio.kafka.KafkaConfig;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
@@ -167,23 +163,6 @@ public class CustomFieldTestBase extends TestBase {
     } catch (IOException | URISyntaxException e) {
       Assert.fail(e.getMessage());
       return null;
-    }
-  }
-
-  public static void updateKafkaConfigField(String fieldName, String newValue) {
-    try {
-      KafkaConfigSingleton instance = KafkaConfigSingleton.INSTANCE;
-      Field kafkaConfigField = KafkaConfigSingleton.class.getDeclaredField("kafkaConfig");
-      kafkaConfigField.setAccessible(true);
-
-      KafkaConfig kafkaConfig = (KafkaConfig) kafkaConfigField.get(instance);
-      Field envIdField = KafkaConfig.class.getDeclaredField(fieldName);
-      envIdField.setAccessible(true);
-      envIdField.set(kafkaConfig, newValue);
-
-      kafkaConfigField.set(instance, kafkaConfig);
-    } catch (NoSuchFieldException | IllegalAccessException e) {
-      e.printStackTrace();
     }
   }
 }
